Move tests file
