+++ _jestsetup.js
@@ -10,7 +10,7 @@ global.chrome.runtime = {
 
 global.fixture = {
   load: file => {
-    const fullPath = path.join(__dirname, '..', file);
+    const fullPath = path.join(__dirname, file);
     const fixture = fs.readFileSync(fullPath);
     document.body.innerHTML = fixture.toString();
   },
