+++ lib/__tests__/click-handler.test.js
@@ -1,7 +1,7 @@
 import assert from 'assert';
 import sinon from 'sinon';
 import $ from 'jquery';
-import clickHandler from '../lib/click-handler';
+import clickHandler from '../click-handler';
 
 describe('click-handler', () => {
   const sandbox = sinon.sandbox.create();
