+++ lib/__tests__/insert-link.test.js
@@ -1,5 +1,5 @@
-import insertLink from '../lib/insert-link.js';
-import { REQUIRE } from '../packages/helper-grammar-regex-collection';
+import insertLink from '../insert-link.js';
+import { REQUIRE } from '../../packages/helper-grammar-regex-collection';
 
 describe('insert-link', () => {
   const DEFAULT_REGEX = /foo ("\w+")/;
