+++ lib/plugins/__tests__/docker.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import dockerImage from '../../lib/plugins/docker';
+import dockerImage from '../docker';
 
 describe('docker-image', () => {
   it('resolves foo to https://hub.docker.com/_/foo', () => {
