+++ lib/plugins/__tests__/dot-net-core.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import dotNetCore from '../../lib/plugins/dot-net-core';
+import dotNetCore from '../dot-net-core';
 
 describe('dotNetCore', () => {
   it('resolves Microsoft.NETCore.App to https://www.nuget.org/packages/Microsoft.NETCore.App', () => {
