+++ lib/plugins/__tests__/dot-net.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import dotNet from '../../lib/plugins/dot-net';
+import dotNet from '../dot-net';
 
 describe('dotNet', () => {
   it('resolves EntityFramework to https://www.nuget.org/packages/EntityFramework', () => {
