+++ lib/plugins/__tests__/go.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import goUniversal from '../../lib/plugins/go';
+import goUniversal from '../go';
 
 describe('go-universal', () => {
   const path = 'octo/dog.go';
