+++ lib/plugins/__tests__/haskell.test.js
@@ -1,6 +1,6 @@
 import assert from 'assert';
-import Haskell from '../../lib/plugins/haskell';
-import githubSearch from '../../lib/resolver/github-search.js';
+import Haskell from '../haskell';
+import githubSearch from '../../resolver/github-search.js';
 
 describe('haskell', () => {
   const path = '/user/repo/blob/d6/lib/plugins/javascript.js';
