+++ lib/plugins/__tests__/homebrew-manifest.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import homebrew from '../../lib/plugins/homebrew-manifest';
+import homebrew from '../homebrew-manifest';
 
 describe('homebrew-file', () => {
   it('resolves a dependency of a homebrew-science formula to both homebrew-science and homebrew-core', () => {
