+++ lib/plugins/__tests__/java.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import Java from '../../lib/plugins/java';
+import Java from '../java';
 
 describe('Java', () => {
   it('resolves java core links', () => {
