+++ lib/plugins/__tests__/javascript.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import plugin from '../../lib/plugins/javascript.js';
+import plugin from '../javascript.js';
 
 describe('javascript-universal', () => {
   it("resolves 'foo/bar.js' like 'foo'", () => {
