+++ lib/plugins/__tests__/less.test.js
@@ -1,6 +1,6 @@
 import assert from 'assert';
-import Less from '../../lib/plugins/less';
-import githubSearch from '../../lib/resolver/github-search.js';
+import Less from '../less';
+import githubSearch from '../../resolver/github-search.js';
 
 describe('Less', () => {
   const path = '/octo/dog.less';
