+++ lib/plugins/__tests__/nodejs-relative-path.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import plugin from '../../lib/plugins/nodejs-relative-path.js';
+import plugin from '../nodejs-relative-path.js';
 
 describe('nodejs-relative-path', () => {
   it('resolves `__dirname + "/views"` relative to the current directory', () => {
