+++ lib/plugins/__tests__/python.test.js
@@ -1,6 +1,6 @@
 import assert from 'assert';
-import python from '../../lib/plugins/python';
-import liveResolverQuery from '../../lib/resolver/live-resolver-query.js';
+import python from '../python';
+import liveResolverQuery from '../../resolver/live-resolver-query.js';
 
 describe('python', () => {
   const path = '/octo/dog.py';
