+++ lib/plugins/__tests__/rust.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import rustCrate from '../../lib/plugins/rust';
+import rustCrate from '../rust';
 
 describe('rust-crate', () => {
   it('resolves hamcrest using the live-resolver', () => {
