+++ lib/plugins/__tests__/sass.test.js
@@ -1,6 +1,6 @@
 import assert from 'assert';
-import Sass from '../../lib/plugins/sass';
-import githubSearch from '../../lib/resolver/github-search.js';
+import Sass from '../sass';
+import githubSearch from '../../resolver/github-search.js';
 
 describe('Sass', () => {
   const path = '/octo/dog.scss';
