+++ lib/plugins/__tests__/vim.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import vimPlugin from '../../lib/plugins/vim';
+import vimPlugin from '../vim';
 
 describe('vim-plugin', () => {
   it('resolves VundleVim/Vundle.vim to https://github.com/VundleVim/Vundle.vim', () => {
