+++ lib/resolver/__tests__/github-search.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import githubSearch from '../../lib/resolver/github-search.js';
+import githubSearch from '../github-search.js';
 
 describe('github-search', () => {
   const path = '/octo/foo.txt';
