+++ lib/resolver/__tests__/relative-file.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import relativeFile from '../../lib/resolver/relative-file.js';
+import relativeFile from '../relative-file.js';
 
 describe('relative-file', () => {
   const path = '/octo/file.txt';
